separate out CLI; start handler implementation
